requirements.txt
Django==1.10.3
django-datatables-view==1.14.0
MySQL-python==1.2.5
argparse==1.2.1
django-mailer==1.2.2
lockfile==0.12.2
mailer==0.8.1
mysqlclient==1.3.9
wsgiref==0.1.2
django-mysql-2.4.1
